Fix error in exclusive_scan_by_segment algorithm for USM device memory (#449)
Fixed crash in exclusive_scan_by_segment algorithm for the case when we use algo with USM device memory and count of items equal to 1. Now in this case we don't copy data on a host from USM device memory in exclusive_scan_by_segment_impl function and call hetero algo implementation.