Address PR comments and minor fixes
This commit primarily addresses review comments. Additionally, the max
local allocation size requirement has been increased by an order to
ensure sufficient SLM for the group algorithms. The last significant
change is that the work group segmented scan returns the known identity
instead of 0 for the work item with the 0th ID.

Signed-off-by: Matthew Michel <matthew.michel@intel.com>