Redefine transform_reduce's scratch & result mem
For L0 backend, it's faster to have a USM host allocation and to write
the reduction result directly to that.

For Nvidia, host USM is expensive, and it's faster to have a single
USM device allocation for both the result and the intermediate scratch
when required.

This commit combines the two approaches into a struct
__usm_host_or_unified_storage, based on the previous
__usm_host_or_buffer_storage.