Adding std::vector<T, usm_allocator>::iterator to is_passed_directly types (#1438)
Changes USM allocator allocated std::vectors from being processed like host-side iterators to being "passed directly" (into sycl kernels) when it is possible to detect them.  The C++ standard library implementation may or may not include allocator information in the iterator type.  If it does, we can detect such iterators, and treat them as "passed directly".  Where it is not possible, they are still treated as host-side iterators and wrapped in a sycl::buffer.
Signed-off-by: Dan Hoeflinger <dan.hoeflinger@intel.com>