[test] Reduce for algorithm testing with debug builds to prevent timeouts (#2058)
This PR constrains testing in SYCL for-based algorithms when compiling in debug mode as test timeouts have been observed on BMG / Arc. With debug builds, a single input to test the large-submitter path is used per type. Release build testing remains unchanged, and a larger set of inputs are assessed.

---------

Signed-off-by: Matthew Michel <matthew.michel@intel.com>